refactored data package to d2data + refactored d2data functions into parent struct
